Move rtl to repo test utils
core/test-utils is discouraged anyway
