+++ packages/material-ui/src/ButtonBase/Ripple.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { assert, expect } from 'chai';
 import { spy, useFakeTimers } from 'sinon';
 import { getClasses } from '@material-ui/core/test-utils';
-import { cleanup, createClientRender } from '@material-ui/core/test-utils/createClientRender';
+import { cleanup, createClientRender } from 'test/utils/createClientRender';
 import TouchRipple from './TouchRipple';
 import Ripple from './Ripple';
